Project/requirements.txt
flask==3.0.3
openai==1.40.3
python-dotenv==1.0.1
